+++ src/app/(sok)/_components/searchBox/SearchBox.jsx
@@ -1,50 +1,88 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { FetchAction, useFetchReducer } from "@/app/_common/hooks/useFetchReducer";
-import * as actions from "@/app/_common/actions";
 import SearchCombobox from "@/app/(sok)/_components/searchBox/SearchCombobox";
-import { getSearchBoxOptions } from "@/app/(sok)/_components/searchBox/buildSearchBoxOptions";
+import { BodyShort, Box, Button, HStack, Link as AkselLink } from "@navikt/ds-react";
+import { DISTANCE, POSTCODE, URL_VERSION } from "@/app/(sok)/_components/searchParamNames";
+import fixLocationName from "@/app/_common/utils/fixLocationName";
+import { InformationSquareIcon, TrashIcon } from "@navikt/aksel-icons";
+import SaveSearchButton, { toSavedSearch } from "@/app/lagrede-sok/_components/SaveSearchButton";
+import useSearchQuery from "@/app/(sok)/_components/SearchQueryProvider";
 
-let suggestionsCache = [];
-const CACHE_MAX_SIZE = 50;
-const MINIMUM_LENGTH = 1;
+function SearchBox({ aggregations, locations, postcodes }) {
+    const searchQuery = useSearchQuery();
 
-function SearchBox({ aggregations, locations }) {
-    const [suggestionsResponse, suggestionsDispatch] = useFetchReducer([]);
-
-    async function fetchSuggestions(value) {
-        const cached = suggestionsCache.find((c) => c.value === value);
-        if (cached) {
-            suggestionsDispatch({ type: FetchAction.RESOLVE, data: cached.data });
-            return;
-        }
-        let data;
-        try {
-            data = await actions.getSuggestions(value, MINIMUM_LENGTH);
-        } catch (err) {
-            // ignore fetch failed errors
-        }
-        if (data) {
-            suggestionsCache = [{ value, data }, ...suggestionsCache].slice(0, CACHE_MAX_SIZE);
-            suggestionsDispatch({ type: FetchAction.RESOLVE, data });
-        }
-    }
-
-    function handleValueChange(value) {
-        if (value && value.length >= MINIMUM_LENGTH) {
-            fetchSuggestions(value);
-        }
-    }
-
-    const allSuggestions = [...suggestionsResponse.data];
+    const drivingDistanceFilterActive =
+        searchQuery.has(POSTCODE) && searchQuery.get(POSTCODE).length === 4 && searchQuery.get(DISTANCE) > 0;
+    const onlyPostcodeOrDistanceFilterActive =
+        searchQuery.size === 2 && (searchQuery.has(POSTCODE) || searchQuery.has(DISTANCE));
+    const savedSearchUrlWithoutVersion = toSavedSearch(searchQuery.urlSearchParams);
+    savedSearchUrlWithoutVersion.delete(URL_VERSION);
+    const showSaveAndResetButton = savedSearchUrlWithoutVersion.size > 0 && !onlyPostcodeOrDistanceFilterActive;
+    const chosenPostcodeCity =
+        drivingDistanceFilterActive && postcodes.find((p) => p.postcode === searchQuery.get(POSTCODE)).city;
 
     return (
-        <section aria-label="Søk etter stilling" className="mb-4">
-            <SearchCombobox
-                onChange={handleValueChange}
-                options={getSearchBoxOptions(aggregations, locations, allSuggestions)}
-            />
-        </section>
+        <Box
+            padding={{ xs: "6 4", md: "6 12" }}
+            background="surface-alt-1-subtle"
+            borderRadius={{ md: "large" }}
+            className="SearchContainer"
+        >
+            <section aria-label="Søk etter stilling" className="mb-4">
+                <SearchCombobox aggregations={aggregations} locations={locations} />
+            </section>
+            {drivingDistanceFilterActive && (
+                <HStack align="center" wrap={false} gap="2">
+                    <HStack gap="2">
+                        <BodyShort weight="semibold">Reisevei:</BodyShort>
+                        <BodyShort>
+                            Innen {searchQuery.get(DISTANCE)} km fra {searchQuery.get(POSTCODE)}{" "}
+                            {fixLocationName(chosenPostcodeCity)}
+                        </BodyShort>
+                    </HStack>
+                    <Button
+                        type="button"
+                        variant="tertiary"
+                        onClick={() => {
+                            searchQuery.remove(POSTCODE);
+                            searchQuery.remove(DISTANCE);
+                        }}
+                        icon={<TrashIcon aria-hidden="true" />}
+                        size="small"
+                    >
+                        Fjern
+                    </Button>
+                </HStack>
+            )}
+            <HStack minHeight="36px" className="mt-3" gap="2" columns="2" align="baseline">
+                <div>
+                    {showSaveAndResetButton && (
+                        <>
+                            <SaveSearchButton size="small" />
+                            <Button
+                                type="button"
+                                variant="tertiary"
+                                onClick={() => {
+                                    searchQuery.reset();
+                                }}
+                                icon={<TrashIcon aria-hidden="true" />}
+                                size="small"
+                            >
+                                Nullstill søk
+                            </Button>
+                        </>
+                    )}
+                </div>
+                <BodyShort>
+                    <AkselLink href="/slik-bruker-du-det-nye-soket">
+                        <span className="link-icon">
+                            <InformationSquareIcon aria-hidden="true" />
+                        </span>
+                        <span>Slik bruker du det nye søket for best resultat</span>
+                    </AkselLink>
+                </BodyShort>
+            </HStack>
+        </Box>
     );
 }
