+++ src/app/lagrede-sok/_components/SaveSearchButton.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from "vitest";
-import { FROM, SEARCH_STRING, SECTOR, SIZE, SORT, VERSION } from "@/app/(sok)/_components/searchParamNames";
+import { FROM, SEARCH_STRING, SECTOR, SIZE, SORT, URL_VERSION } from "@/app/(sok)/_components/searchParamNames";
 import { toSavedSearch } from "@/app/lagrede-sok/_components/SaveSearchButton";
 
 describe("test toSavedSearch", () => {
@@ -24,11 +24,11 @@ describe("test toSavedSearch", () => {
     test("should contain version parameter", () => {
         const input = new URLSearchParams();
         input.append(SEARCH_STRING, "react");
-        input.append(VERSION, "1");
+        input.append(URL_VERSION, "1");
 
         const expected = new URLSearchParams();
         expected.append(SEARCH_STRING, "react");
-        expected.append(VERSION, "1");
+        expected.append(URL_VERSION, "1");
 
         const result = toSavedSearch(input);
